eink/requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cycler==0.10.0
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.3
numpy==1.19.4
pandas==1.1.5
Pillow==8.0.1
pkg-resources==0.0.0
py-sds011==0.9
pyparsing==2.4.7
pyserial==3.5
python-aqi==0.6.1
python-dateutil==2.8.1
pytz==2020.5
requests==2.25.1
RPi.GPIO==0.7.0
scipy==1.5.4
seaborn==0.11.1
six==1.15.0
spidev==3.5
urllib3==1.26.2
waveshare-epd==0.0.0
Werkzeug==1.0.1
